Extend generator by a method for executing sub-processes
When executing sub-processes in Java, it is important to constantly
read from the process' output and error streams as those have limited
space. If a buffer is full, the sub-process blocks until the data is read.
This could lead to a deadlock. The newly added method constantly forwards
the process streams to other streams.

See #130