C++: forbid usage of arrays in C++
This adds a check forbidding any use of arrays in C++. The change also includes
some unit tests. There is no sensible way for plain arrays to be safely used in
the genrated C++ code. With this change we enforce the usage of `std::array` or
`std::vector` instead of plain arrays.

See #136