C++: move "using namespace" declarations back to the top of the file
The previous solution did not work, as for instance, directives in the
constructor needed the namespaces to be available as well. This change solves
the compilation issue, but potentially inserts "using namespace" in header
files. We could circumvent this issue in the future, by placing each reactor
in a namespace.