Added back the pthread_join because it is essential to the federation working correctly
Sometimes, a downstream federate can exit without waiting for the upstream federate to finish. This could cause a send_message or send_timed_message to fail catastrophically (sometimes during message delivery). The graceful TCP shutdown demands that all incoming messages be received until the sender closes the socket. In case of a loop, there should not be a deadlock because federates close their outbound connections unconditionally during termination before waiting for incoming connections to close. If the program hangs at this line, it could most likely be due to an error somewhere else in the program logic.