Fix sudo permissions (#430)
Add `sudo` to the nightly workflow. That was not required in the local build (probably a quirk of the `act` tool) but is necessary here.

Also sorry for the PR for something small like this, but `master` is protected. Which is good practice but always annoying when doing tinkering with CI. 😅