Add InvalidSourceException, fix tests
The test previously showed a wrong error message,
because expected and actual value were inverted,
and because the caught exception was not the
exception we expect. Now we throw a more specific
exception type and only catch that one. This way
if the test fails the error message will be reasonable.

Throwing `new Exception` from xtend code is anyway extremely bad
practice, because the exception cannot be caught from a Java caller.
Xtend sneaky-throws.