don't reset the error reporter
Resetting the error reporter in doGenerate has the side-effect of deleting
all messages reported earlier in the validator. This leads to warnings from the
validator not being reported by lfc. Since the only place where
we need the reset method is for clearing markers in the IDE, this change
introduces a separate method only for the EclipseErrorReporter precisely
to clean IDE markers.

See #522