Optimize broadcast connections with after
Given a broadcast connection with N ports on the left an M ports on the right,
we currently instantiate M delay reactors and broadcast to the delays. However,
it would be more efficient to only instantiate N delay reactors and broadcast
the output of the delays. This change implements this optimization and also
adds a new testcase.