Allow using type parameters in ctor params
Remove need for Clone bound on parameters.
This also means sometimes clone will need to
be called explicitly, this should be rare though.

Remove the params struct field. Parameters need
to be stored inside state variables to be accessible
within reactions. This fixes the meaning of a parameter
to be a "constructor parameter", not an implicit field
stored somewhere.