Fix grammar
Replace Value members with different
node types that implement the interface Value.
Uniformize treatment of Value everywhere in grammar.

This is still going to need some work to
streamline the generators. I feel like
everything around values has been duplicated
10 times everywhere... The Rust branch
introduces a couple of things that may be
used to refactor other generators more gradually
and purge GeneratorBase, so let's wait for that.