Disable chain ID
The current chain ID assignment and usage is incorrect and will lead to nondeterminism in the exectuion of the program. Until that is fixed, we need to disable the usage of chain IDs. This is what was causing the PingPongDistributed example/benchmark to fail. See #608. However, the problem is not limited to federated execution.