Fix C tests
Note that now, the syntax to assign a parameter (in instantiations)
is the same as the syntax for any initializer. This is the case in
the entire grammar and in all targets.

Hence in

    reactor Foo(p(...))

the `...` mean exactly the same as in

    new Foo(p(...))

Notably, this is not necessarily the same thing as

    reactor Foo(p = ...)

where again, the `...` mean the same as in

    new Foo(p = ...)

Hence in LF-C, since the syntax to initialize a list is `name: type(a,b,c)`,
the syntax to override a list parameter is `new Foo(name(a,b,c))`.
And since in LF-C, `name: type = (a,b,c)` is forbidden (it's a tuple literal),
it stands to reason that `new Foo(name = (a,b,c))` is now forbidden.

To alleviate this we could allow `(a,b,c)` to be interpreted as an
array initializer in LF-C instead of forbidding it entirely.
But I think this is not satisfactory as that syntax means something
else in other targets, and C uses braces to denote array initializers...
Maybe we should replace the tuple syntax in C with brace initializers.