Simplify the big benchmark and extend file comment
This change simplifies the big benchmark and brings it closer to the Akka
implementation in the savina suite. The most important change is the
introduction of separate ports for ping and pong. This eliminates some weirdness
from before, where it was not possible to send a worker a ping and a pong
messages simultaneously.

The benchmark still performs very bad compared to Akka (about 10x slower). I
investigated this issue and added an explanation to the file comment.