Remove disconnected network ports
Keep track of triggers of network reactions of a federate if those triggers are from remote federates. Remove them in code generators after levels are assigned because these triggers reference remote federates and might require structures that will not be generated for the current federate. There is a fixme for the TS target.