Address benchmark test failure.
Note that this changes the promised behavior of a method. Although it would be possible to fix the problem without changing the method's promised behavior, I did this because I suspect that the best way to eliminate bugs like this is to minimize the number of special cases.

The issue was that reactors with multiple multiports were handled incorrectly.