LSP: Break #line directives a little bit.
I think this is okay because line directives are on their way out. The problem that this solves (involving CodeMap generation when a code block has leading blank lines) could have been fixed without breaking #line directives, but this way the number of lines (and special cases!) in our code base decreases.