Rust: Ask cargo what the package root is.
This patches a tiny edge case that might not show up in practice. The problem is that the Rust code is being generated into an existing project (i.e., there is already a Cargo.toml file in the src-gen directory), which means that sometimes srcGenPkgPath != the actual package root (according to cargo). The cost of collecting and parsing the metadata is probably not negligible, so it would be nice if we didn't have to do this.

See https://github.com/rust-lang/cargo/issues/4933.