LFCommand: Try more aggressively to execute.
The only fully general way that I have found to check if something is executable is to try to start the process and see if an `IOException` occurs. `File.canExecute()` is insufficient -- it can return `true` when the answer is really `false`. Furthermore, successful termination of the `where` command on Windows apparently does not guarantee that the right executable will be used.