LSP tests: Make always different but locally replicable.
This is a response to failures on remote that could not be reproduced locally. It is unclear what causes it, but if it is nondeterminism in order of test execution, then the previous way of fixing the seed was no good. With this approach, the seed will be different for every test run, but at least the seed will be reported so that failures can be debugged.