diagram: Add categories explizitly to the synthesis options.
Since the lsp does not allow to send cyclic structures via json the
category of a synthesis option is only retrievable by its id. If the
category is not also added as synthesis-option, klighd-vscode omits all
options for which the category is not retrievable. Adding the categories
explicitly solves this.