[formatting] Make badness minimizer work properly.
The time and conceptual effort required for this particular commit
reflect poorly on the implementation. This was partly because complex
logic was implemented in the toString function, which makes debugging
exceedingly difficult when setting breakpoints (since the debugger
itself invokes toString). However, the complexity also comes from the
statefulness of the renderable components (a possible design mistake),
as well as from the (possibly inherent) difficulty of keeping track of
comments as them bubble up to ever higher syntactic levels in search
of a home.