[formatting] Aesthetic adjustment.
That this aesthetic improvement should have to be done in this way is
concerning. Ideally, it would be performed by changing the badness
function, but the badness function itself already could in principle
do the job in some cases -- but in the implementation, it doesn't,
for an algorithm that with certainty minimizes global badness would
be have exponential time complexity.

That said, this is a sensible and probably reasonably future-proof
heuristic. Programming constructs tend to have bodies at the end
(or lists of expressions, or whatever) and metadata at the beginning
(types, mutability, etc.), so usually, the expression at the end is
the largest, and the consequences of representing it differently are
the greatest (in particular, in terms of reducing line length
violations).