Do not minimize the shadow jar
Issue #1284 was caused by the mimization step of the shadowJar gradle plugin.
The main problem is, that we don't import kotlin classes directly, but load
them manually in order to work around the build problems in Eclipse... Hence,
the minimizer thinks it can exclude the Kotlin classes. While in
principle it is possible to add particular excludes, this does not work properly
when we compile a second time. The reason for this is unclear, but we might as
well disable the minimization. This slightly increases the jar size from 23 MB to
28 MB.

Closes #1284