[formatting] Associate comments to ancestors.
Comments already have been associated with ancestors, but this makes
it possible for that to happen even if there are intervening
non-whitespace AST nodes. The issue is that we have a policy of
putting comments sometimes at the end of the first line of the
representation of an AST node, and in that case, we do not want the
comment to become associated with child nodes that are also on that
line.