Suppress failing Arduino tests.
This follows up on a conversation with @arengarajan99 and Marten about
the Arduino tests. The rationale is that soon, our Arduino support will
be entirely different anyway, and that we can tolerate a temporary
regression in the interim as a shortcut to save person-hours.

Obviously I take responsibility if we have to scramble to get support
back in time for a release.