Fix NPE in Epoch.
Recently, I fixed a bug in Windows error reporting in which the drive
prefix (e.g., C:) was being excluded from the path. I suspect that the
reason why this caused NPEs in Epoch is that Epoch uses special URIs
that are not file system paths, even in local development.