Concentrated PowerShell commands
Right now, launch.ps1 will, just like how lfc and lff (which are symbolic links) on *NIX works, get the name of the invoker and run the corresponding jar file, which is not as PowerShell-y. We can use arguments like outlined here: https://stackoverflow.com/a/43940404.
Nevertheless, it appears it does work somehow.