Avoid reading the same input stream twice
This fixes a nasty bug in FileUtil, where an input stream was read twice when
copying. The first read was to check if the destination file needs to be
modified, and the second read was to actually perform the copy. Since input
streams can only be read once, the second read would then copy an empty byte
array and hence write an empty file.

This likely fixes https://github.com/lf-lang/lingua-franca/issues/1460