Get gemm example working.
The dgemm example is not too pretty unfortunately. What is the
idiomatic way to fan out a request? In Java, you would spawn the tasks
and then block on each of their futures, one at a time, but that does
not work in LF because if responses do not come back at the same logical
time, then you have to save the ones that come early in order to process
them all at once. And if you process them in the order of arrival, then
it is like CSP -- even if the implementation is deterministic, the model
that you are using to predict what will happen is nondeterministic, and
that's not good.