Try to address failing CCpp tests.
They were failing because of linker errors. Usually when CCpp fails with
linker errors and C does not, it is because of the shenanigans that we
used to do with our #includes. So I merged the latest changes from
reactor-c/main. This introduced another problem, so I cherry-picked
Erling's commit related to the LF_THREADED compile definition.