Try to pass TS serialization test.
As far as CI is concerned I suspect that it might have been sufficient
merely to correct the path (".resolve("src")"), but on my machine I also
had to install google-protobuf probably because it was not already
installed globally. This could have been done by modifying the produced
package.json instead of just copying it, as is done in
updatePackageConfig, but the method used in this commit also seems to
work.