Address deadlock in LoopDistributedCentralizedPhysicalAction.
This is an off-by-one error in the computation of
WORKERS_NEEDED_FOR_FEDERATE. The code in this commit is copied from
other code from master/CUtil which was last touched by Steven Wong and
which is deleted in this branch.

I should also point out that the previous commit, which updated the
reactor-c version, probably would not have been necessary given this
commit.