Do not cache the targetConfig.
This matters in case of AST transformations. The test framework uses an
AST transformation to set tests as CCpp, and it needs the context set in
order to do the AST transformation. But the context would need to be
computed after the AST transformation in order for it to get the right
target. This is the least invasive way to sidestep this dependency
cycle. I acknowledge that there may be performance consequences, but I'm
not concerned about performance consequences until they are shown to be
of practical significance. The fact is that if we do not cache, we do
not have two inconsistent views of the targetConfig in the presence of
AST transformations that touch the targetConfig.