Address federated Python LSP test failure.
The issue was not exactly a concurrency bug in the sense that data was
not being set to an invalid state (AFAICT) by concurrent access. But it
was in the sense that shared state was causing an error in the
compilation of one federate to abort the compilation of another
federate. This is OK -- it might even be desirable -- but it is a source
of complexity.

The root cause of the problem is that non-errors were being reported as
errors. That is the only aspect of the issue that is being fixed here,
and maybe the only aspect that really needs to be fixed.