Print a useful error message if the runtime sources are missing
We used to have an error message in lfc, that notified the user if the runtime
files are missing (i.e. the submodules are not checked out). It seems that with
the introduction of `FileUtil` this feature got lost. It currently only
complains if a resource cannot be found at all, but not if the resource is an
empty directory. Aside from a bit of cleanup, this fix ensures that a useful
error message is also produced if a submodule is missing.