Fix passing of command line options from lfc to the generator
Currently, all command line options that need to be passed on from lfc to the
code generator do not work. This change provides a fix. Concretely, it replaces
the automagic passing in a for loop from before with explicit code for each
individual option. While it is a bit more verbose, this has some advantages:
- The code is readable and understandable
- IntelliJ produces a useful warning when an option is not used and therefore
not passed on

This change further removes some unused code and includes other small cleanups.
All in all, this removes all warnings shown bu IntelliJ.