Update ts-tests.yml.
When we make changes to reactor-ts and accompanying changes in
lingua-franca we cannot point to the version of reactor-ts that is
published on NPM. We need to clone a specific ref and point to it.

Locally, pointing to the correct runtime version fixes the problem that
I observed in #1607. If this change to the yaml file is correct then
this should make that PR mergeable.