Do not report warnings as errors.
This does not change the output of `lfc` even when an error appears in
an imported reactor. The code just was not doing what it was claiming to
be doing AFAICT other than redundant work, since it was being invoked
after a round of validation checks had already passed. And it is even
worse than that, because it was running the validator without the right
ValidationIssueAcceptor injected, resulting in behavior that was not
consistent with that implemented in StandaloneIssueAcceptor. I decided
to delete the whole thing rather than a) make this code work properly
and b) spend some unknown number of hours fighting with Guice. The fact
is that it is doubtful whether the behavior that it was supposed to have
is even what we want. I am not sure that users really need their import
statements to be highlighted with errors reported on the imported
reactors.