Disable the "dubious ownership" thing.
See
https://github.com/git/git/commit/8959555cee7ec045958f9b6dd62e541affb7e7d9.
We are not accessing repositories on a shared drive that we do not
control. I don't see how this security issue could apply to us.

I am changing to a * from what Erling wrote 4 days ago because I was
still having issues running CI from reactor-c, and the error message I
got told me to do something different:
git config --global --add safe.directory /__w/reactor-c/reactor-c
Let's not play whack-a-mole with the directories that we need to mark as
safe.