Try to pass unit tests.
The error gets reported with an unexpected line number (not even in the
right file) when a mistake is inserted into the type for the list of
time values. I do not want to fix this right now (honestly it is the C++
compiler's fault, I do not even know how we would fix this), so instead
I will try to cover it up.