Bugfix.
The problem here is that we use the uniqueID of the federate both before
and after we transform it into separate programs. After we transform it,
the name of the instance depends on the name of the main reactor, which
might be different (indeed, it must be different in order for us to
eliminate the possibility of name collisions with other parts of the
program). So this is a bit of a hack, but I'm not sure what else to do.