Address federated test failures.
It is bad that in order to make this change it is necessary to edit code
in several places. But it isn't clear how to fix this because different
parts of the code operate on the program at different stages in the
compilation process, either before or after it has been split into many
programs.