Do not doubly mark as federated.
This is hacky. The problem is that the generated code does not have a
federated reactor, but we do an AST transformation so that we can
pretend that it is (I do not want to know why). So we end up running the
code generator on a federated program, even though it is supposed to
result on code for a single federate. So if the federate had its name
prefixed once in the main run of the code generator, its name will be
prefixed again in the nested run of the code generator.