Do not use hashcodes as keys.
There are a few reasons for this change
* the key is everywhere being computed from an Instantiation, so this
 change reduces the total amount of code
* hash codes can occasionally collide, and I was not sure that this
 particular hashcode was very effective at minimizing collision
 probability.
* but more importantly, the hash codes were being computed from
 information that does not uniquely identify the Instantiation. It was
 being computed from the name of the variable that holds the instance
 and from the name of the reactor class being instantiated. So if the
 same reactor class were instantiated and stored in a variable of the
 same name in multiple places, this could result in bugs.

This change means that we are using physical equality to check the keys
to the hashmap because we don't have our own IR and we can't override
the hashcode and equals functions of Instantiation. But I think that's
OK. After AST transformations, which should happen as upstream as
possible, we are working with one single object representation of the
program.