Fix line adjustment logic for federated programs.
It is bad that I had to modify ToLf to do this since the logic isn't
really in ToLf. The problem was that the line adjustment relies on
tagging the target code (without the fat braces), and the target code is
not recorded separately from the fat braces if they are entered into the
MalleableString machinery as all one string.