Handle shutdown.
This terminates the spawned formatter by closing its input stream. The
formatter should detect that the stream is closed and understand that it
will not receive any more files to format and exit gracefully. This
prevents the formatter from outliving its Gradle daemon.

If the daemon is terminated forcefully e.g. with a -9 SIGTERM rather
than a -15 SIGKILL, then this will not work. Otherwise I believe that
this should release the resource (get rid of the spawned formatter)
correctly.