Work around build dependency issue.
The problem was not that Gradle runs any spotlessLinguaFranca tasks as
part of its other operations, but that it configures the task before
starting any other operation (including clean). The solution is to start
the formatter lazily -- to wait until spotlessLinguaFranca actually
operates on a file before trying to start up the formatter process.