Fix warnings reported on CliBase.
The comparison of strings to string literals looks like a critical
error which I think would have been caught by sufficient tests.

One of these warnings is incorrect, but we might as well silence it with
an assertion.

The warning about toArray is merely a matter of best practice. A human
couldn't be expected to know about this, but the warning emitted by
IntelliJ is correct.