Use a flag to indicate successful exit.
For some reason, creating a trap for SIGTERM results in SIGTERM not
causing the program to exit.

In any case, we really do not want the program to exit without killing
the RTI and federates under any circumstance unless we are sure that the
program terminated successfully. Therefore, it seems safest to require
positive confirmation via an EXITED_SUCCESSFULLY flag that the program
really did exit successfully.