Make name mangling stateless again.
The assignment of mangled names is too complex when compiling several
distinct programs in the same JAR execution because it causes them to
interact with each other by changing the number of reactors that have
the same name. We can patch this up, but it is safer and easier to just
eliminate the global variables. It will make the generated code uglier,
but maybe life is too short to worry about that.