Make formatter check itself at runtime.
This may slow down the formatter. It also is not certain to work
because isEqual could contain mistakes. However, it may substantially
reduce the likelihood that the formatter silently produces wrong
results. It is hard to be sure about this, however, because it is
already rare for the formatter to produce wrong results.