Update formatter.
This runs another pass of optimization. Yes, this is
totally a hacky heuristic, but it should not affect idempotency nor
other correctness criteria. This can affect performance, but I modified
the formatter to keep track of whether anything has changed so that it
knows whether further optimization passes have the potential to give any
improved results.