Make level the max of the levels depended on.
This is the same as the lub/sup and makes the method match the spec. It
is also a potential performance optimization because it exposes more
parallelism with less synchronization in the case of parallel paths with
different numbers of reactions.
However it does mean that the number of reactions in the current level
can change after we have already started executing the level. I think
that we already had that problem due to some details surrounding
federated execution, but I am not sure.