Update submodule.
Do not decrement the reference count twice.
In my most recent meeting with @jackykwok2024 I believe we determined
that although there are multiple memory leaks, the most severe one is
fixed in the commit being referenced here, and that in any case the
other commit in reactor-c is not less leaky than the one referenced
here. Besides, I have found that in at least one test program a segfault
is fixed by using the commit referenced here (d28a9f5).