Mark SpuriousDependency as passing.
It was necessary to make a modification to SpuriousDependency in order
for it to pass because the original version had top-level reactions and
state variables. We do not currently have a good way to get top-level
reactions and state variables to work properly. The problem is that is
we simply put the reactions where we think they "belong" and replicate
the state variables, then the values of the state variables will not be
the same across different federates. This is a can of worms that is
outside the scope of this PR to fix.