Try again to work around the test failures.
I noticed that the test that fails most often had an out of memory error
in CI at one point. Usually the problem was something else --
undetermined because the debug output was being truncated. However this
is an indication that the debug output is excessive. We have spent a lot
of time struggling with excessive debug output in CI and so I am
resorting to this most radical solution of allowing to disable it for
specific tests.