Fix IllegalArgumentException in diagram synthesis.
This is a runtime type error -- this is my bad, as such things should not
happen in well written Java code. This commit modifies FormattingUtil
to make it slightly more type-safe. It also changes the diagram
synthesis to bypass the regular formatting logic -- the formatting logic
seems more heavyweight than necessary since in diagrams we should not
usually have to worry about how to represent comments and where to break
lines (comment representation is the reason why the target was needed by
the FormattingUtil.render function).