Record execution times.
This is in response to a timeout. I would like to get an idea of the
range of normal variability in the execution times of these tests in GH
actions. If a test fails with a timeout but finishes quickly when
re-run, then that is an indication that the timeout might be a real
deadlock.