Fixed trimming of the recorded test output
This fixes a bug in our test framework. It seems like
`Runtime.getRuntime().freeMemory()` does not indicate the total memory that can
still be allocated, but only the free memory that is already allocated by the
runtime. Therefore, the recorded output was often cut prematurely.
This fix only considers the size of the recorded string buffer. It also makes
sure, that the buffer is trimmed after deleting elements.