Be extra explicit with include guards.
I believe that this is only a problem for Arduino because Arduino brings
in header files that are not actually #included. This solution is a
little bit of a hack because it should not be necessary to add this
extra #ifndef, but it solves the problem and it should be mostly
harmless for users that do not target Arduino.