Move to same category as other CCpp tests.
I think this is OK; this test is inherently both a generics test and a
CCpp test (because the special characters come from using C++), so it
seems fine to put it in either category.

But it should be in the target category so that it is not run on
Windows.