Report unsupported target properties as errors
Currently, if the target declarationu uses a property that is not
supported by the target, this is reported as a warning. This is
problematic, as we and potentially users don't notice if a target
property is not supported anymore. For instance, we had several uses of
the `flags` property in the C tests although it was renamed. Also the
benchmarks used properties that we removed (See https://github.com/lf-lang/benchmarks-lingua-franca/pull/61).

This PR converts the warning message into an error, adjusts the unit
tests, and fixes some of the C tests. I opted for simply dropping the
`flags` property, as apparently it wasn't required and not used for a while.
In particular, some tests attempted to avoid optimization, which is not
needed as we compile the tests in Debug mode (whithout optimization)
anyway.