Fixed null pointer exception in message reporting
This fixes https://github.com/lf-lang/lingua-franca/issues/2228. When a message
is reported on an eObject, but the object is null, this let to a null pointer
exception. This change falls back to reporting nowhere, if the given object is
null.

This generalizes and replaces the fix in https://github.com/lf-lang/lingua-franca/pull/2229.