Fix another unrelated bug uncovered by testing fix
Note how non-obvious it is that this is a bug or even why the fix works.
This is a consequence of reliance on implicit reliance on big chunks of
shared mutable state. The resource is modified; the context is
computed from the resource; therefore, the resource has to be modified
before the context is computed.