Give up trying to uninstall GCC
This is what happens in CI:
> Run sudo rm -f $(which gcc)
rm: /usr/bin/gcc: Read-only file system

This seems to indicate that there is no way to move or delete the GCC
binary, and it will not work to change the PATH to not include /usr/bin
because that will break a lot of things. Therefore, there probably is
no easy way to make GCC invisible on the system, so if we want a quick
fix, then we should give up exercising that code path in CI. The code
path under test is already difficult to maintain anyway and works on a
"best effort" basis.